+++ CHANGES.rst
@@ -11,7 +11,9 @@ Versioning <https://semver.org/spec/v2.0.0.html>`__.
 
 Unreleased
 ----------
+
 - Improved documentation of the raster modules.
+- Validate bounds of input raster arrays to be within WGS84 lat/lon coordinates.
 
 
 0.17.0 - 2023-06-27